Updated openstack/openstack Project: openstack/tempest  197055992861e7c9d1deaad55659639dbd7ba2ae
Fix disk_available_least in JSON schemas

The hypervisor JSON schemas were not consistent with the value for
disk_available_least. One allowed for 'null' values, but the other
two did not. This allows 'null' in all schemas (Ironic does not
set this value so it will be 'null' when the compute api tests
are run).

Closes-Bug: #1337416

Change-Id: I71e6c5f0800f7641c49d8058596831f011328ec2