Updated openstack/openstack Project: openstack/nova  7b68766e2c66e784b578227b245cc446d27f5a1b
virt: use context & instance as param names in migrate APIs

The vast majority of methods in the ComputeDriver use
'context' and 'instance' as parameter names. The only
exceptions are the migration related methods which use
'ctx' and 'instance_ref'. Change the latter to be
consistently named with the rest of the file

Partial-bug: #1333219
Change-Id: I0457e3e70ee58a9cbfc69f8cfe675de0504cc781