Updated openstack/openstack Project: openstack/horizon  83c550946819f4cbea16a9aee99e5acbef3804b2
Raise sane error when failing to modify security group

Horizon catches and propagates recoverable exceptions back to the
user.  This code previously caught all exceptions and then raised
Exception (with an unhelpful message), which Horizon propagated as an
even less helpful message.  By raising the error as a
novaclient.exception.ClientException, we give Horizon a fighting
chance of passing a useful (but sanitized) error message to the user.

This changes the message the user sees from:

Error: There was an error submitting the form. Please try again.

To:

Error: Unable to modify instance "test10".
Error: Failed to modify 1 instance security groups

The error condition can be simulated by trying to modify security
groups on a suspended instance.

Closes-Bug: #1329888
Change-Id: I38d9977d6d20dfeddbb77b1a7e63e74597bf7331