Updated openstack/openstack Project: openstack/nova  96959a69996f5aace8504fc5ee8d6f6d747001bd
Change instance disappeared during destroy from Error to Warning

Currently Nova logs an error if a libvirt domain disappears during
while waiting for it to be destroyed, but the code actually treats
this (correctly) as a recoverable situation since the end result
is the required one. Hence this should be logged as a warning not
an error.

Change-Id: Ie72cd11e22c5f049d3263e58bec286914798147a
Closes-bug: #1336767