Updated openstack/openstack Project: openstack-dev/devstack  3876d77d92298cbbdfe1bec1f0efa462bee947c4
Avoid huge libvirtd.log files on Ubuntu hosts

The libvirt log filter settings match against the filename of
the libvirt source emitting the log message. Normally these
file names are relative to the source tree root, but in the
Ubuntu binary packages these have somehow ended up as absolute
filenames from the OS root. This means that a log filter of
'1:libvirt' which is only intended to match src/libvirt.c
will in fact match every single file. This caused enourmous
log files on Ubuntu hosts running the gate.

The fix is to use '1:libvirt.c' as a more specific filename
match, but we can't do this unconditionally because libvirt
>= 1.2.3 does not use filenames for log filter matching
anymore. So only change the match on Ubuntu hosts for now,
since that's where the original problem lies.

While doing this, also turn off the logging of object ref
and unref operations, since those pollute the logs with lots
of noise.

Change-Id: I71b67507a4e68a7bff0c358857aaaac08ef0c420