Updated openstack/openstack Project: openstack/cinder  1449571ef0673330a0b48076ea003630b58b1874
sync periodic_task fix from incubator

Set the value of time the periodic task was last run to a multiple of
the spacing value rather than the actual last time run. This ensures
tasks runs regularly but avoids synchronizing that interval with other
nodes.

This patch also removes the coalescing of tasks which are _nearby_ (0.2
seconds away).

Changes:

periodic_task
-------------
4dbd3aa Make periodic tasks run on regular spacing interval
21249f1 Fix parenthesis typo misunderstanding in periodic_task

Change-Id: I6c06ae73b18d4425a987a215a050bc70335e490e