Updated openstack/openstack Project: openstack/oslo.rootwrap  fb8f8af47f106205c64713fcffe05ac5e9397157
Continue on failure of leaf filters of chaining filters

Currently wrapper.py has an implementation error in chaining filters, which
is expecting match_filter() returns None when no leaf filter matches to
arguments, although it actually raises an exception.

This patch fixes this and adds an unit test for it.

Closes-Bug: #1340792
Change-Id: I1ac4535ad176d78769c497fa037285bc610acadf