Updated openstack/openstack Project: openstack/cinder  cc2cad87c6300da282d3c04610bab7dbc99941d2
Remove $sqlite_db from default database connection

sqlite_db used to be a config option in the DEFAULT section. This
allowed it to be used in config template substitution. Now that it is in
the database section, the template substitution fails with a
NoSuchOptError.

The quick fix is to just hardcode the default connection option to refer
directly to 'cinder.sqlite' instead.

Closes bug: 1339295

Change-Id: Ifd2267aaa8c55d78f88ce6648badefe3f9f664ee