Updated openstack/openstack Project: openstack/python-keystoneclient  3b73bcf98fa44029a438bf1cd316b35db6161305
Direct move of the revoke model from keystone server

This commit copies the revoke model from Keystone to
keystoneclient. The only difference in the model.py
file between keystoneclient and keystone is the
import due to a change of the oslo-incubator path
(project name is keystoneclient not keystone).

Blueprint: revocation-event-api

Change-Id: I6215d6679f9bb0dff6941c2a65ceeefd1ff9d88a