Updated openstack/openstack Project: openstack/nova  64a5e566cfe1067c59a747e8da692e59067ffd84
Prevent max_count > 1 and specified ip address as input

Previously, if one did:
   nova boot --num-instances 2 --nic net-id=<net_id>,v4-fixed-ip=10.0.0.17
this would result in all of the instances failing to boot because neither
nova-network or neutron allows multiple instances to have the same ip_address.
This patch fixes this from occuring by checking for this and raising an error
to the caller from nova-api.

Change-Id: I7e44d58e66aeb10fb15c3553072f988f65ab823b
Closes-bug: 1318754