Updated openstack/openstack Project: openstack/requirements  b49b1ac881248ecf7641cbd57c34377f6aa5fad3
Add requests-mock package to global requirements

requests-mock is a mocking layer for the requests package. It is
Apache-2 licensed.

The httpretty library has been used to provide a mocking layer for HTTP
requests. This has generally been successful however there have been a
number of problems with it over our usage, including poor python 3
support and a number of backwards incompatible releases from the
maintainer. It is also difficult to package by distributions.

httpretty was particularly chosen because it was compatible with both
httplib and requests, this is no longer such a requirement as almost all
OpenStack libraries have now moved to requests. Given that we can move
to a library with less complexity and more inline with OpenStack tools.

It is written in such a way that it is fairly simple to convert existing
httpretty test suites or httmock test suites to requests-mock.

I am the author of requests-mock and am also happy to move it to
stackforge, or otherwise connect it to gerrit so as we will not have any
future incompatible issues.

More info:
 - https://github.com/jamielennox/requests-mock
 - http://requests-mock.readthedocs.org/en/latest/

Change-Id: I09f2a22f84752571a278e9b90901ecb1fbedb654