Updated openstack/openstack Project: openstack/python-keystoneclient  0e8673769929026b9a3facc078c876be82cdf7cb
Pass roles manager to user manager

This is part of breaking a cyclical dependency where the client has a
reference to the manager but the manager has a reference to the client.

To be able to pass in an adapter to the managers we need to remove the
client specific usages of manager, including referring to other managers
via the client.

Change-Id: I8fba7ece05e33d519eb31994a072275598ba76ea