Updated openstack/openstack Project: openstack/python-keystoneclient  658b6900493c9b169dc0dc971e8a9a0f087a0df5
Pass user and roles manager to tenant manager

This is part of breaking a cyclical dependency where the client has a
reference to the manager but the manager has a reference to the client.

To be able to pass in an adapter to the managers we need to remove the
client specific usages of manager, including referring to other managers
via the client.

Change-Id: Ie238dda046ec8bbdaca749a691df3da63f5a9a63