Updated openstack/openstack Project: openstack/nova  d1f9af71521acc70bde34e2f75c9985a9da14723
Add host_ip to compute node object

This adds host_ip to the compute_node object as an IPAddress
typed field. The property is already present in the database
and is required to enable use of the compute_node object in
the resource tracker.

This is part of: blueprint make-resource-tracker-use-objects

Change-Id: Id238538a51b5a7a8014fc991def7817959904a95