Updated openstack/openstack Project: openstack/nova  7dc2fd715704690db42e57d25d90519636e93337
Fix interfaces template for two interfaces and IPv6

If you have more than one interface and IPv6 enabled, then the interfaces file
emitted would have the incorrect syntax: the `auto eth1` line would be
appended to the end of the `gateway` line of the previous stanza.

This was a result of improper whitespace control in the Jinja template which
this patch fixes.

In order to prevent regressions, this patch swaps out the existing tests with
corresponding tests that are whitespace sensitive. This is important because
it's difficult to tweak Jinja templates in a way that fixes the whitespace
for one scenario while not breaking another.

Closes-Bug: 1339888

Change-Id: Id5566ea2f93ee1a6b8bc42e83bc64a1c6a09e680