Updated openstack/openstack Project: openstack/tempest  96646364daa4cc43011e7e29d6ab4c41e0b38fbf
First part of moving success response checking to identity client

Adds client success checking and removes test code checking of responses
for methods contained in the IdentityClient classes. This includes
IdentityClientJSON, IdentityClientXML, IdentityV3ClientJSON and
IdentityV3ClientXML. It also includes v3 ServiceClientJSON and
ServiceClientXML since the service methods seem to have been split out when
the v3 clients were created. This is a large patch but it was important to
make these changes together because there were changes in return codes between
v2 and v3 but the method names are in many cases the same. I didn't want to
risk a mistake by trying to separate v2/v3 into separate changes.

The doc for some keystone list apis says they can return 200 or 203. Only
one of the calls to list-users-for-tenant was checking for both so it seems
that 203 is not actually returned. I removed the check for 203.

Doing this uncovered a case where there was an unbound variable and an
incorrect status check that combined to create a test that passed. This bug was
also fixed.

Closes-Bug: 1336826
Partially implements: blueprint client-checks-success

Change-Id: I75633b5b94b1750a623f6a2b8adaae1efc5c7c9c