Updated openstack/openstack Project: openstack/trove  f383f7e28d08b24c2b42e6c2b79c1c78432f5b14
Change default for update_status_on_fail

Default 'update_status_on_fail' to True so that instances do not get stuck
in the BUILD state.

When this option is set, instances that remain in BUILD for longer than
the datastore's configured 'usage_timeout' will be considered failed and
moved to the ERROR state.

Closes-Bug: 1339826
Change-Id: I5f4b734830a21a1fbb6f169edb5424bac6f883cd