Updated openstack/openstack Project: openstack/horizon  63f770ea79225dd7926658605bc0c27c8ef5635a
Remove requirements style check

This is incorrect and should never have been added.  The order of
lines in requirements files affects the pip dependency resolver.
Therefore, while it's nice to add new packages in alphabetical
order, it must not be required.

Move pbr to the top of the list as it should always be installed
first.  Move hacking to the top of the test-requires list.

Change-Id: I81f7e75b77baccd12d9660dc009f30f2f2c891bc