Updated openstack/openstack Project: openstack/nova  89ebdfd9b7a22fe75fd74a27b232fce6a74752ef
libvirt: fix typo in fakelibvirt listAllDomains()

It looks like this code isn't used in the tests at all since we replace
it with a fake in the list_instance_domains() tests ... so what's the
point of it?

We could also add tests for the function in test_fakelibvirt, but that
feels rather pointless too?

Change-Id: I8fa5301ad86e2eb117f40d26a7cfe7b9d86c873a