Updated openstack/openstack Project: openstack-dev/devstack  f29569db7a1de1f15b654e1db131de249143e210
Add pre-install phase for extras.d plugins

An additional call hook for the extras.d plugins that is called before
any service installation occurs turns out to be needed by Ceph and
useful for other plugins.

This is called between the installation of the system packages listed
as prerequisites and the installation of the actual services.

Change-Id: Id723bdc7542182d2b2eff40c0dc0127307f094c5