Updated openstack/openstack Project: openstack/cinder  c15d2640f49d6a3c82c3135e40bc3b61f96f1752
Clean up base Volume Driver

This is a preliminary cleanup of the base
Volume Driver to prepare for adding new
architecture which keeps the Control calls
and connection portions of the drivers
seperated from each other.

This doesn't break existing compatability or drivers
but just makes some of the necessary tweaks and additions
to enable implementation of decoupled Control and Data
sides of the Volume drivers.

POC for this proposal is here:
https://review.openstack.org/#/c/104701/

FYI check-up2-date is going to be the death of us all!

Change-Id: Ifc844c49ee9ff1580d5da0879e7d499bed72dac1