Updated openstack/openstack Project: openstack/python-keystoneclient  a58bb4b27ada086f850ee990da9ed0ef15f3294c
Allow passing kwargs from managers to session

Resource managers may need to pass certain overrides down to the session
layer. The most common example would be to allow them to specify a
different interface to issue a request to than is common. It is possible
later that we may wish to expose this ability to users as well so that
they can choose in there own programs what interface, endpoint name etc
a particular request should be issued to.

Change-Id: I7ee0931d4ea78cf9f463b2f0b54457226c5bea8d