Updated openstack/openstack Project: openstack/horizon  db24daad58766c681c37666987ea7a6ce005ffbd
Adding the data_image_registry panel for Sahara

The data_image_registry panel allows for tagging
images with the flavors and versions of hadoop that they
support.  Those tags are used in creating clusters to
be sure that a cluster has all the necessary pieces to
function properly.
This code was originally from:
https://github.com/openstack/sahara-dashboard

Change-Id: I647c03a0ce9b802dd1100bce49dc46e24cac7190
Partial-Implements: blueprint merge-sahara-dashboard
Co-Authored-By: Nikita Konovalov <nkonovalov@mirantis.com>
Co-Authored-By: Dmitry Mescheryakov <dmescheryakov@mirantis.com>