Updated openstack/openstack Project: openstack/tempest  e40ef9d6026200770e9a789f5016e4103ade6743
fix javelin to let you specify imgdir

The biggest issue with javelin right now is locating images on
disk. This needs a reference to a base devstack directory for it
to work in a generic way. This will let us specify the devstack
base dir on the cli, and provide a well understood dir under it
in the yaml to support images.

Disable volumes until we can bootstrap the code.

Change-Id: If5f2674179eea4972ad43a1581073d483bb75d79