Updated openstack/openstack Project: openstack/horizon  301d48757ace57f8da8633d3ff7ab17878732bce
Add datastore and version information for backups

We should display the datastore name and datastore
version associated with a backup in the backups
table and the backup details view.

There is already a blueprint to add this information
to the instances table and details view so we have
used the same code and formatting here to ensure
consistency.

Change-Id: I286d430d4d670ee07298495e8f1b78795fa64609
Closes-Bug: #1340366